+++ tests/testthat/helper-model.R
@@ -30,8 +30,8 @@ h_get_general_model <- function() {
 
 h_get_model <- function() {
   .Model(
-    dose = function(prob, param1) {
-      prob
+    dose = function(x, param1) {
+      x
     },
     prob = function(dose, param1) {
       dose